fix(rating-system): code refactor

+++ rating-system/src/main/java/ru/walkername/rating_system/events/RatingDeleted.java
@@ -2,34 +2,34 @@ package ru.walkername.rating_system.events;
 
 public class RatingDeleted {
 
-    private int userId;
+    private Long userId;
 
-    private int movieId;
+    private Long movieId;
 
     private int rating;
 
     public RatingDeleted() {
     }
 
-    public RatingDeleted(int userId, int movieId, int rating) {
+    public RatingDeleted(Long userId, Long movieId, int rating) {
         this.userId = userId;
         this.movieId = movieId;
         this.rating = rating;
     }
 
-    public int getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public int getMovieId() {
+    public Long getMovieId() {
         return movieId;
     }
 
-    public void setMovieId(int movieId) {
+    public void setMovieId(Long movieId) {
         this.movieId = movieId;
     }
